requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
idna==2.7
Jinja2==2.10
MarkupSafe==1.0
python-craigslist==1.0.6
python-dotenv==0.9.1
requests==2.20.0
six==1.11.0
urllib3==1.24
